unireserve-backend/requirements/base.txt
django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
celery
redis
django-celery-beat
django-ratelimit
pillow
django-filter
drf-spectacular
PyMySQL
